move login beforeEach to cypress support file
