+++ cypress/support/index.js
@@ -14,7 +14,16 @@
 // ***********************************************************
 
 // Import commands.js using ES2015 syntax:
-import './commands'
+import "./commands";
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
+before(function() {
+  function login() {
+    cy.get("input[name=username]").type("admin");
+    cy.get("input[name=password]").type("password");
+    cy.get("button[id=login-submit]").click();
+  }
+  cy.visit("/login");
+  login();
+});
